mqtt_master/requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
psycopg2==2.8.5
requests==2.22.0
urllib3==1.25.6
